Restructured the backend
