+++ backend/views.py
@@ -1,0 +1,11 @@
+from app import app
+from models import Product, Transaction
+from blueprints.inventory import inventory_bp
+from blueprints.product import product_bp
+from blueprints.sales import sales_bp
+
+
+# Register the blueprints
+app.register_blueprint(inventory_bp, url_prefix="/inventory")
+app.register_blueprint(product_bp, url_prefix="/product")
+app.register_blueprint(sales_bp, url_prefix="/sales")
